Use u64 for network byte counts instead of i64
- Changed rx_speed and tx_speed properties to u64
- Removed lossy as i64 cast; u64 is the semantically correct type
- QML handles u64 (Qt ulonglong) transparently

+++ src/bridge.rs
@@ -3,8 +3,8 @@ pub mod qobject {
     extern "RustQt" {
         #[qobject]
         #[qml_element]
-        #[qproperty(i64, rx_speed)]
-        #[qproperty(i64, tx_speed)]
+        #[qproperty(u64, rx_speed)]
+        #[qproperty(u64, tx_speed)]
         type NetworkMonitor = super::NetworkMonitorRust;
 
         #[qinvokable]
@@ -21,8 +21,8 @@ static NETWORKS: OnceLock<Mutex<Networks>> = OnceLock::new();
 
 #[derive(Default)]
 pub struct NetworkMonitorRust {
-    rx_speed: i64,
-    tx_speed: i64,
+    rx_speed: u64,
+    tx_speed: u64,
 }
 
 impl qobject::NetworkMonitor {
@@ -35,7 +35,7 @@ impl qobject::NetworkMonitor {
                 let iface = default_iface();
                 net.iter()
                     .find(|(name, _)| **name == iface)
-                    .map(|(_, data)| (data.received() as i64, data.transmitted() as i64))
+                    .map(|(_, data)| (data.received(), data.transmitted()))
                     .unwrap_or((0, 0))
             }
             Err(_) => (0, 0),
